requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.3
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.2.9
django-cors-headers==3.10.1
django-extensions==3.1.5
django-organizations==2.0.1
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
mysqlclient==2.0.3
Pillow==8.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==21.0.0
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
zope.interface==5.4.0
